Fix time measurement
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@22506 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ marcel/examples/threads/utils.h
@@ -22,8 +22,8 @@ typedef union u_tick
 
 	struct
 	{
-		unsigned long low;
-		unsigned long high;
+		unsigned low;
+		unsigned high;
 	} sub;
 	
 	struct timeval timev;
@@ -100,7 +100,7 @@ void timing_init(void)
 	GET_TICK(last_event);
 }
 
-double tick2usec(long long t)
+double tick2usec(unsigned long long t)
 {
 	return (double)(t)*scale;
 }
